- Added an ESlint tool to notify me of errors so i could reduce the time debuging wih codacy since codcy uses the same linter - Added the babel core to be able to build to production qucikly - Added the package-lock file to .gitIgnore - Modified some linting rule to make it work with my spacing choice - Updated the readme to display the results of Codacy
