+++ README.md
@@ -1,3 +1,5 @@
 # Quick mechanic
 
 Team-291-Backend
+
+[![Codacy Badge](https://api.codacy.com/project/badge/Grade/0f431f9d67854c3ea59f3010d7f7b383)](https://app.codacy.com/gh/BuildForSDGCohort2/Team-291-Backend?utm_source=github.com&utm_medium=referral&utm_content=BuildForSDGCohort2/Team-291-Backend&utm_campaign=Badge_Grade_Settings)
