+++ package.json
@@ -5,7 +5,8 @@
 	"main": "index.js",
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
-		"start": "nodemon ./index.js --exec babel-node -e js"
+		"start": "nodemon src/index.js --exec babel-node -e js",
+		"build": "babel src -d build"
 	},
 	"repository": {
 		"type": "git",
@@ -19,9 +20,16 @@
 	},
 	"homepage": "https://github.com/Dru-Go/Team-291-Backend#readme",
 	"devDependencies": {
+		"@babel/core": "^7.11.4",
 		"babel-cli": "^6.26.0",
 		"babel-preset-env": "^1.7.0",
-		"babel-preset-stage-0": "^6.24.1"
+		"babel-preset-stage-0": "^6.24.1",
+		"eslint": "^7.7.0",
+		"eslint-config-standard": "^14.1.1",
+		"eslint-plugin-import": "^2.22.0",
+		"eslint-plugin-node": "^11.1.0",
+		"eslint-plugin-promise": "^4.2.1",
+		"eslint-plugin-standard": "^4.0.1"
 	},
 	"dependencies": {
 		"express": "^4.17.1",
